fix: dando acesso ao user no test de deleção

+++ back/app/Http/Controllers/ReservationController.php
@@ -66,8 +66,13 @@ class ReservationController extends Controller
     public function destroy(Request $request, int $reservation)
     {
         try {
-            Log::info('Recebida requisição para deletar reserva do usuário: ' . $request->user()->id);
-            return $this->service->delete($reservation);
+            if(Auth::user()->hasPermissionTo('cancelar reserva')) {
+                Log::info('Recebida requisição para deletar reserva do usuário: ' . $request->user()->id);
+                return $this->service->delete($reservation);
+            } else {
+                Log::info('Usuário sem permissão para deletar reserva');
+                return response()->json(['message' => 'Unauthorized'], 401);
+            }
         } catch (\Throwable $th) {
             Log::error('Erro ao deletar reserva: ' . $th->getMessage());
             return response()->json(['message' => 'Error'], 500);
